2dcontext/drawing-text-to-the-canvas/2d.text.measure.offscreencanvas.actualBoundingBox.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.text.measure.actualBoundingBox</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("/fonts/CanvasTest.ttf");
}
</style>
<body class="show_output">

<h1>2d.text.measure.offscreencanvas.actualBoundingBox</h1>
<p class="desc">Testing actualBoundingBox for OffscreenCanvas</p>


<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>

<ul id="d"></ul>
<script>
var t = async_test("Testing actualBoundingBox for OffscreenCanvas");
test(function () {
  deferTest();
  var f = new FontFace("CanvasTest", "/fonts/CanvasTest.ttf");
  document.fonts.add(f);
  document.fonts.ready.then(() => {
    step_timeout(t.step_func_done(function () {
      var offscreen = new OffscreenCanvas(100, 50);
      var ctx = offscreen.getContext('2d');
      ctx.font = '50px CanvasTest';
      ctx.direction = 'ltr';
      ctx.align = 'left';
      ctx.baseline = 'alphabetic';
      // Some platforms may return '-0'.
      _assertSame(Math.abs(ctx.measureText('A').actualBoundingBoxLeft), 0, "Math.abs(ctx.measureText('A').actualBoundingBoxLeft)", "0");
      // Different platforms may render text slightly different.
      _assert(ctx.measureText('A').actualBoundingBoxRight >= 50, "ctx.measureText('A').actualBoundingBoxRight >= 50");
      _assert(ctx.measureText('A').actualBoundingBoxAscent >= 35, "ctx.measureText('A').actualBoundingBoxAscent >= 35");
      _assertSame(Math.abs(ctx.measureText('A').actualBoundingBoxDescent), 0, "Math.abs(ctx.measureText('A').actualBoundingBoxDescent)", "0");

      _assertSame(Math.abs(ctx.measureText('ABCD').actualBoundingBoxLeft), 0, "Math.abs(ctx.measureText('ABCD').actualBoundingBoxLeft)", "0");
      _assert(ctx.measureText('ABCD').actualBoundingBoxRight >= 200, "ctx.measureText('ABCD').actualBoundingBoxRight >= 200");
      _assert(ctx.measureText('ABCD').actualBoundingBoxAscent >= 85, "ctx.measureText('ABCD').actualBoundingBoxAscent >= 85");
      _assert(ctx.measureText('ABCD').actualBoundingBoxDescent >= 37, "ctx.measureText('ABCD').actualBoundingBoxDescent >= 37");
    }), 0);
  });
});
</script>
</body>>
